+++ NEWS.md
@@ -3,6 +3,9 @@
 - `CreateSeuratObject.Assay` sets Assay key when not present (#29)
 - Ignore warnings when creating an `Assa` from a data frame (#32)
 
+## Added
+- New `CheckMatrix` generic for validating expression matrices
+
 # SeuratObject 4.0.3
 
 ## Changed